+++ CHANGELOG.md
@@ -79,3 +79,47 @@ the following form: YYYY.0M.0D.
 - Designed a StockPortfolio component
 - Designed a BankAccount component
 - Designed a DNASequence component
+
+## [2024.10.02]
+
+### Added
+
+- Designed a proof of concept for BankAccount component
+
+### Updated
+
+- Changed design to include multiple features such as adding accounts,
+removing accounts, adding and removing money, calculating interest, getting
+the type of an account, getting values of accounts, adding a overdraft fee,
+getting the total number of accounts, and getting the total money in all
+accounts.
+
+## [2024.10.17]
+
+### Added
+
+- Designed kernel and enhanced interfaces for BankAccount component
+
+### Updated
+
+- Changed design to include all method contracts for the implementations of all
+the method from the proof of concept
+- Extended Standard into BankAccountKernel and included 4 methods in the
+kernel
+- Extended BankAccountKernel into BankAccount and included 6 more methods in
+the BankAccount
+
+## [2024.10.30]
+
+### Added
+
+- Designed abstract class for BankAccountSecondary component
+
+### Updated
+
+- Changed design to include Secondary method with the data of a HashMap with
+Strings as the keys and Account class as the values as long as data for the
+size of the map or the total accounts
+- Added Code for totalAccounts, addAccount, removeAccount, and hasAccount
+methods and implemented them in the Secondary method
+- Added implementation for hashCode and toString for BankAccount